+++ src/main/java/kr/co/springbootweeklymission/wallet/domain/repository/JdbcWalletRepository.java
@@ -1,7 +1,7 @@
 package kr.co.springbootweeklymission.wallet.domain.repository;
 
-import kr.co.springbootweeklymission.infrastructure.error.exception.DuplicatedException;
-import kr.co.springbootweeklymission.infrastructure.error.model.ResponseStatus;
+import kr.co.springbootweeklymission.common.error.exception.DuplicatedException;
+import kr.co.springbootweeklymission.common.response.ResponseStatus;
 import kr.co.springbootweeklymission.member.domain.entity.Member;
 import kr.co.springbootweeklymission.member.domain.repository.JdbcMemberRepository;
 import kr.co.springbootweeklymission.voucher.domain.entity.Voucher;
